+++ src/app/app.module.ts
@@ -3,21 +3,21 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header.component';
-import { HomeComponent } from './home/home.component';
-import { ConverterPanelComponent } from './converter-panel/converter-panel.component';
+import { HeaderComponent } from './shared/components/header/header.component';
+import { HomeComponent } from './pages/home/home.component';
+import { ConverterPanelComponent } from './shared/components/converter-panel/converter-panel.component';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatOptionModule } from '@angular/material/core';
 import { FormsModule } from '@angular/forms';
-import { CurrencyCardComponent } from './currency-card/currency-card.component';
+import { CurrencyCardComponent } from './shared/components/currency-card/currency-card.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatIconModule } from '@angular/material/icon';
-import { DetailsComponent } from './details/details.component';
-import { CurrencyService } from './services/currency.service';
+import { DetailsComponent } from './pages/details/details.component';
+import { CurrencyService } from './shared/services/currency/currency.service';
 import { HttpClientModule } from '@angular/common/http';
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { ChartComponent } from './chart/chart.component';
+import { ChartComponent } from './shared/components/chart/chart.component';
 import { NgChartsModule } from 'ng2-charts';
 
 @NgModule({
