+++ src/app/pages/details/details.component.html
