+++ src/app/pages/details/details.component.scss
