+++ src/app/pages/details/details.component.spec.ts
