+++ src/app/shared/components/chart/chart.component.scss
