+++ src/app/shared/components/converter-panel/converter-panel.component.spec.ts
