+++ src/app/shared/components/currency-card/currency-card.component.scss
