+++ src/app/shared/components/currency-card/currency-card.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CurrencyCardComponent } from './currency-card.component';
-import { ConvertedValue } from '../models/currency.model';
+import { ConvertedValue } from '../../../models/currency.model';
 
 describe('CurrencyCardComponent', () => {
   let component: CurrencyCardComponent;
