+++ src/app/shared/components/header/header.component.scss
